requirements.txt
fastapi
uvicorn
catboost
pydantic
pandas
matplotlib
torch
scikit-learn
joblib
python-multipart
seaborn
statsmodels
numpy
kafka-python